Controls updated
Binds should be able to be referenced elsewhere so as to make the basic interactions bindable via settings
Currently the default keys act as placeholders to the binds within the if loop. Unsure if will actually work, but the keys are definitely bound. 

May be better to change to an "active()" function and circumvent the loop and direct reference to default keys to ensure it does work when we make the settings modifiers and/or have a splash screen requring player to pick keys at start which will then be saved to locale and be referencable.